pyproject.toml
[project]
name = "bomb-audio-new-stack"
version = "0.1.0"
description = "Bomb audio detection with TensorFlow 2.19 + AutoKeras 2.2.0 stack"
requires-python = ">=3.11,<3.12"

dependencies = [
    # Core scientific packages
    "numpy",
    "scipy",
    "matplotlib",
    "pandas",
    "tqdm",
    "scikit-learn",
    "h5py",
    # Audio processing
    "librosa",
    "pydub",
    "soundfile",
    "audioread",
    # Additional ML dependencies
    "joblib",
    "resampy",
    # Audio augmentation
    "audiomentations",
    # Additional dependencies for AutoKeras 2.2+
    "dm-tree",
    "optree",
    "ml-dtypes",
    "namex",
    "rich",
    "markdown-it-py",
    "mdurl",
    "typing-extensions",
    # Jupyter
    "jupyter",
    "ipykernel",
    "tensorflow-macos>=2.16.0",
    "tensorflow-metal>=0.8.0",
    "autokeras>=2.0.0",
    "keras>=3.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "flake8",
]
override-dependencies = [
    "tensorflow-text==2.19.0"  # Override AutoKeras pin to use ARM64-compatible version
] 
